+++ ada-malef/src-base/malef-surfaces.ads
@@ -7,7 +7,7 @@
 --                                  S P E C                                  --
 --                                                                           --
 -------------------------------------------------------------------------------
---     Copyright (c) 2020 José Antonio Verde Jiménez All Rights Reserved     --
+--  Copyright (c) 2020-2021 José Antonio Verde Jiménez  All Rights Reserved  --
 -------------------------------------------------------------------------------
 -- This file is part of Malef.                                               --
 --                                                                           --
@@ -26,7 +26,6 @@
 --                                                                           --
 -------------------------------------------------------------------------------
 
-with Malef.Colors;
 private with Ada.Finalization;
 
 --